requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
streamlit
requests
python-dotenv
pydantic
pyjwt
pydantic-settings
